actions: Use super-linter from actions repository
The local action of `super-linter` was lackluster because it was
missing several modifications that we found to be suitable for our
content. The actions repository that we developed for the Security
Summer School contains all these updates and also provides a unified
approach to lint multiple repositories without duplicating
configurations unnecessarily.

Signed-off-by: Teodor Dutu <teodor.dutu@gmail.com>